backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyJWT==2.9.0
bcrypt==4.2.0
email-validator==2.2.0
scikit-learn==1.5.2
joblib==1.4.2

